requirements-render.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# OpenAI Whisper
openai-whisper==20231117

# PyTorch CPU only (smaller)
torch==2.2.0
torchaudio==2.2.0
--index-url https://download.pytorch.org/whl/cpu

# Audio/Video Processing
ffmpeg-python==0.2.0
numpy==1.24.3
soundfile==0.12.1

# Production server
gunicorn==21.2.0